CI 工作流 + ruff 全量治理
- 新增 .github/workflows/ci.yml: push/PR 自动跑 ruff check + pytest
- ruff 安全/机械修复 800+ 项 (导入排序/未用导入/pygrade 等), 469 测试全绿
- 修复 3 个真问题: human_ai_gateway 漏导 Literal(F821)、
  privacy/temporal worker asyncio 悬挂任务未持有引用(RUF006)、
  contract_service 无效分支(RUF034)
- pyproject 固化风格豁免: 中文全角字符/camelCase API/内存单例可变默认
- API 层异常补 from None 防内部错误链泄露

+++ production/backend/app/services/bank_statement_parser.py
@@ -12,10 +12,11 @@ import re
 from datetime import datetime
 
 from app.schemas.parsers import (
-    BankStatementParseResult, BankStatementRecord, OcrRequest,
+    BankStatementParseResult,
+    BankStatementRecord,
+    OcrRequest,
 )
 from app.services.ocr_service import OcrService, ocr_service
-
 
 # ============================================================================
 # 内置 seed 样本 (3 家银行)
@@ -349,7 +350,7 @@ def _parse_amount_cents(line: str) -> int:
         return 0
     amt_str = m.group(1).replace(",", "").replace("+", "")
     try:
-        return int(round(float(amt_str) * 100))
+        return round(float(amt_str) * 100)
     except ValueError:
         return 0
 
@@ -434,9 +435,7 @@ class BankStatementParserService:
             )
             if bal_match:
                 try:
-                    balance_cents_after = int(
-                        round(float(bal_match.group(1).replace(",", "")) * 100)
-                    )
+                    balance_cents_after = round(float(bal_match.group(1).replace(",", "")) * 100)
                 except ValueError:
                     pass
